moving utilities.py into utils folder
